test(statedb): move extract_block_number_and_index_number test code
