+++ crates/store/src/state_db.rs
@@ -129,6 +129,15 @@ impl CheckPoint {
         })
     }
 
+    #[cfg(test)]
+    pub fn do_extract_block_number_and_index_number(
+        &self,
+        db: &StoreTransaction,
+        db_mode: StateDBMode,
+    ) -> Result<(u64, u32), Error> {
+        self.extract_block_number_and_index_number(db, db_mode)
+    }
+
     fn extract_block_number_and_index_number(
         &self,
         db: &StoreTransaction,
@@ -539,111 +548,5 @@ impl<'a, 'db> CodeStore for StateTree<'a, 'db> {
             Some(slice) => Some(Bytes::from(slice.to_vec())),
             None => None,
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{CheckPoint, StateDBMode, SubState, WriteContext};
-
-    use crate::{traits::KVStore, transaction::WithdrawalReceipt, Store};
-
-    use gw_common::merkle_utils::calculate_state_checkpoint;
-    use gw_db::schema::COLUMN_INDEX;
-    use gw_types::{
-        packed::{
-            AccountMerkleState, Byte32, GlobalState, L2Block, L2BlockCommittedInfo, L2Transaction,
-            RawL2Block, TxReceipt, WithdrawalRequest,
-        },
-        prelude::{Builder, Entity, Pack, PackVec, Unpack},
-    };
-
-    #[test]
-    fn checkpoint_extract_block_number_and_index_number() {
-        let store = Store::open_tmp().unwrap();
-        let block_store = store.begin_transaction();
-
-        let default_state_checkpoint: Byte32 = {
-            let post_state = AccountMerkleState::default();
-            let root: [u8; 32] = post_state.merkle_root().unpack();
-            let checkpoint: [u8; 32] =
-                calculate_state_checkpoint(&root.into(), post_state.count().unpack()).into();
-            checkpoint.pack()
-        };
-
-        let raw_block = RawL2Block::new_builder()
-            .state_checkpoint_list(vec![default_state_checkpoint; 5].pack())
-            .build();
-
-        let block = L2Block::new_builder()
-            .transactions(vec![L2Transaction::default(); 2].pack())
-            .withdrawals(vec![WithdrawalRequest::default(); 3].pack())
-            .raw(raw_block)
-            .build();
-
-        let block_number = block.raw().number().unpack();
-        let block_hash = block.raw().hash();
-        assert_eq!(0u64, block_number);
-
-        block_store
-            .insert_block(
-                block.clone(),
-                L2BlockCommittedInfo::default(),
-                GlobalState::default(),
-                vec![TxReceipt::default(); 2],
-                vec![WithdrawalReceipt::default(); 3],
-                Vec::new(),
-            )
-            .unwrap();
-        block_store
-            .insert_raw(COLUMN_INDEX, 0u64.pack().as_slice(), &block_hash)
-            .unwrap();
-        block_store.commit().unwrap();
-
-        let db = store.begin_transaction();
-
-        let checkpoint = CheckPoint::from_genesis();
-        let maybe_block_idx =
-            checkpoint.extract_block_number_and_index_number(&db, StateDBMode::Genesis);
-        assert!(maybe_block_idx.is_ok());
-        assert_eq!(maybe_block_idx.unwrap(), (0, 0));
-
-        let checkpoint = CheckPoint::new(0, SubState::Withdrawal(1));
-        let maybe_block_idx =
-            checkpoint.extract_block_number_and_index_number(&db, StateDBMode::ReadOnly);
-        assert!(maybe_block_idx.is_ok());
-        assert_eq!(maybe_block_idx.unwrap(), (0, 1));
-
-        let checkpoint = CheckPoint::new(0, SubState::Tx(1));
-        let maybe_block_idx =
-            checkpoint.extract_block_number_and_index_number(&db, StateDBMode::ReadOnly);
-        assert!(maybe_block_idx.is_ok());
-        assert_eq!(maybe_block_idx.unwrap(), (0, 4));
-
-        let checkpoint = CheckPoint::new(1, SubState::Block);
-        let maybe_block_idx =
-            checkpoint.extract_block_number_and_index_number(&db, StateDBMode::ReadOnly);
-        assert!(maybe_block_idx.is_ok());
-        assert_eq!(maybe_block_idx.unwrap(), (1, 0));
-
-        let checkpoint = CheckPoint::new(1, SubState::Tx(1));
-        let maybe_block_idx =
-            checkpoint.extract_block_number_and_index_number(&db, StateDBMode::ReadOnly);
-        assert_eq!(
-            maybe_block_idx.unwrap_err().to_string(),
-            "DB error can't find block hash"
-        );
-
-        let checkpoint = CheckPoint::new(1, SubState::Withdrawal(1));
-        let maybe_block_idx = checkpoint
-            .extract_block_number_and_index_number(&db, StateDBMode::Write(WriteContext::new(2)));
-        assert!(maybe_block_idx.is_ok());
-        assert_eq!(maybe_block_idx.unwrap(), (1, 1));
-
-        let checkpoint = CheckPoint::new(1, SubState::Tx(1));
-        let maybe_block_idx = checkpoint
-            .extract_block_number_and_index_number(&db, StateDBMode::Write(WriteContext::new(2)));
-        assert!(maybe_block_idx.is_ok());
-        assert_eq!(maybe_block_idx.unwrap(), (1, 3));
     }
 }
